% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emimr.R
\name{emimr}
\alias{emimr}
\title{Correlation}
\usage{
emimr(expression, methylation, miRNA)
}
\arguments{
\item{expression}{expression data}

\item{methylation}{methylation data}

\item{miRNA}{miRNA data}
}
\value{
result
}
\description{
Correlation
}
