% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emimr_count.R
\name{emimr_count}
\alias{emimr_count}
\title{Count the genes}
\usage{
emimr_count(emimr_result)
}
\arguments{
\item{emimr_result}{emimr function result dataframe name}
}
\value{
count table
}
\description{
Count the genes
}
