% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMImR_import.R
\name{EMImR_import}
\alias{EMImR_import}
\title{Import files for the EMImR function}
\usage{
EMImR_import(path)
}
\arguments{
\item{path}{file that contains data csv files}
}
\value{
expression, methylation and miRNA
}
\description{
Import files for the EMImR function
}
