% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF_import.R
\name{GSDF_import}
\alias{GSDF_import}
\title{import files for GSDF function}
\usage{
GSDF_import(path)
}
\arguments{
\item{path}{file that contains the data csv files}
}
\value{
expression, methylation,  miRNA, DEGs , DMGs and GDEIRs
}
\description{
import files for GSDF function
}
